pyproject.toml
[tool.poetry]
name = "shirley"
version = "0.1.10"
description = "It is just doing some stuff intelligently."
authors = ["luojiahai <luo@jiahai.co>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
transformers = "4.36.0"
numpy = "1.26.4"
gradio = "^4.39.0"
pypdf = "^4.3.1"
pytest = "^8.3.1"
tiktoken = "^0.7.0"
einops = "^0.8.0"
transformers-stream-generator = "^0.0.5"
azure-cognitiveservices-speech = "^1.38.0"

[tool.poetry.scripts]
webui = "shirley.webui:main"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
